feat(ui): set root color via ThemeProvider, reorganize theme files structure
