+++ packages/ui/src/design-tokens/index.ts
@@ -1,7 +1,6 @@
 export { sx } from './sx.css';
 export type { Sx } from './sx.css';
-export type { Theme } from './theme.css';
-export { vars } from './theme.css';
-export { ThemeColorScheme, ThemeProvider } from './theme.context';
+export type { Theme } from './theme';
+export { ThemeColorScheme, ThemeProvider, useTheme, vars } from './theme';
 export { style, createVar, globalStyle } from '@vanilla-extract/css';
 export * from '@vanilla-extract/css-utils';
