+++ packages/ui/src/design-tokens/sx.css.ts
@@ -1,6 +1,6 @@
 import { defineProperties, createSprinkles } from '@vanilla-extract/sprinkles';
 
-import { vars } from './theme.css';
+import { vars } from './theme';
 
 const responsiveProperties = defineProperties({
   conditions: {
