+++ packages/ui/src/design-tokens/theme/dark-theme.css.ts
@@ -5,22 +5,24 @@ import {
   darkColorScheme,
   laceGradient,
   lightColorScheme,
-} from './colors.data';
-import { opacities } from './opacities.data';
-import { radius } from './radius.data';
-import { spacing } from './spacing.data';
-import { vars } from './theme.css';
+} from '../colors.data';
+import { opacities } from '../opacities.data';
+import { radius } from '../radius.data';
+import { spacing } from '../spacing.data';
 import {
   fontWeights,
   fontSizes,
   lineHeights,
   fontFamily,
-} from './typography.data';
+} from '../typography.data';
 
-import type { Colors } from './colors.data';
-import type { Elevation } from './elevation.data';
+import { vars } from './theme-contract.css';
+
+import type { Colors } from '../colors.data';
+import type { Elevation } from '../elevation.data';
 
 const colors: Colors = {
+  $lace_typography_main_color: darkColorScheme.$primary_white,
   $lace_gradient: laceGradient,
   $transparent: colorTransparent,
