+++ packages/ui/src/design-tokens/theme/index.ts
@@ -1,0 +1,4 @@
+export { ThemeColorScheme, useTheme } from './context';
+export type { Theme } from './theme-contract.css';
+export { vars } from './theme-contract.css';
+export { ThemeProvider } from './theme-provider.component';
