+++ packages/ui/src/design-tokens/theme/theme-contract.css.ts
@@ -1,16 +1,16 @@
 import { createThemeContract } from '@vanilla-extract/css';
 
-import { colors } from './colors.data';
-import { elevation } from './elevation.data';
-import { opacities } from './opacities.data';
-import { radius } from './radius.data';
-import { spacing } from './spacing.data';
+import { colors } from '../colors.data';
+import { elevation } from '../elevation.data';
+import { opacities } from '../opacities.data';
+import { radius } from '../radius.data';
+import { spacing } from '../spacing.data';
 import {
   fontWeights,
   fontSizes,
   lineHeights,
   fontFamily,
-} from './typography.data';
+} from '../typography.data';
 
 export const vars = createThemeContract({
   spacing,
