+++ packages/ui/src/design-tokens/theme/theme-provider.css.ts
@@ -1,0 +1,7 @@
+import { style } from '@vanilla-extract/css';
+
+import { vars } from './theme-contract.css';
+
+export const root = style({
+  color: vars.colors.$lace_typography_main_color,
+});
